Great cleanup, step 1 - Getting rid of #if 0/#if 1 and ifdefs triggering on JUNK, JNK, IGNORE_ZERO_DURATION, DX_DRAW_BOX_AROUND_TEXT, USE_COCOA_BOTLEFT, NO_CLICK_THRU

+++ src/libambulant/gui/cocoa/cocoa_image.mm
@@ -102,7 +102,7 @@ cocoa_image_renderer::redraw_body(const rect &dirty, gui_window *window)
 			if (opacity < 0.5) {
 				lib::color_t chromakey = ri->get_chromakey();
 				lib::color_t tolerance = ri->get_chromakeytolerance();
-				float components[8] = {
+				CGFloat components[8] = {
 					redc(chromakey)-redc(tolerance), redc(chromakey)+redc(tolerance),
 					greenc(chromakey)-greenc(tolerance), greenc(chromakey)+greenc(tolerance),
 					bluec(chromakey)-bluec(tolerance), bluec(chromakey)+bluec(tolerance),
